api/index.py
from fastapi import FastAPI, HTTPException, Depends
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from pydantic import BaseModel
import jwt
import os
import requests
import json
from datetime import datetime
from typing import Optional, List, Dict, Any

# Configuration
SECRET_KEY = os.getenv("SECRET_KEY", "autonomous-ai-agent-secret-key-2024")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")

# Initialize FastAPI app
app = FastAPI(
    title="Autonomous AI Agent API", 
    version="1.0.0",
    description="Lightweight AI agent for coding tasks"
)

# Authentication
security = HTTPBearer()

def verify_token(credentials: HTTPAuthorizationCredentials = Depends(security)):
    try:
        jwt.decode(credentials.credentials, SECRET_KEY, algorithms=["HS256"])
        return True
    except:
        raise HTTPException(status_code=401, detail="Invalid token")

# Request models
class CodeRequest(BaseModel):
    prompt: str
    language: str = "python"
    max_tokens: Optional[int] = 500

class SearchRequest(BaseModel):
    query: str
    max_results: Optional[int] = 5

class ReasonRequest(BaseModel):
    problem: str
    domain: Optional[str] = "coding"

# Simple template-based code generation
def generate_code_template(prompt: str, language: str) -> str:
    """Generate code using templates and patterns"""
    
    prompt_lower = prompt.lower()
    
    # Common code patterns
    if "fibonacci" in prompt_lower:
        if language == "python":
            return """def fibonacci(n):
    if n <= 1:
        return n
    return fibonacci(n-1) + fibonacci(n-2)

# Example usage
print(fibonacci(10))"""
        elif language == "javascript":
            return """function fibonacci(n) {
    if (n <= 1) return n;
    return fibonacci(n-1) + fibonacci(n-2);
}

// Example usage
console.log(fibonacci(10));"""
    
    elif "binary search" in prompt_lower:
        if language == "python":
            return """def binary_search(arr, target):
    left, right = 0, len(arr) - 1
    
    while left <= right:
        mid = (left + right) // 2
        if arr[mid] == target:
            return mid
        elif arr[mid] < target:
            left = mid + 1
        else:
            right = mid - 1
    
    return -1

# Example usage
arr = [1, 3, 5, 7, 9, 11]
print(binary_search(arr, 7))"""
    
    elif "hello world" in prompt_lower or "print" in prompt_lower:
        if language == "python":
            return f'print("Hello, World!")\n# {prompt}'
        elif language == "javascript":
            return f'console.log("Hello, World!");\n// {prompt}'
        elif language == "java":
            return f'''public class HelloWorld {{
    public static void main(String[] args) {{
        System.out.println("Hello, World!");
        // {prompt}
    }}
}}'''
    
    elif "function" in prompt_lower or "method" in prompt_lower:
        if language == "python":
            return f"""def my_function():
    '''
    {prompt}
    '''
    # TODO: Implement your logic here
    pass

# Example usage
my_function()"""
        elif language == "javascript":
            return f"""function myFunction() {{
    /*
    {prompt}
    */
    // TODO: Implement your logic here
}}

// Example usage
myFunction();"""
    
    # Default template
    if language == "python":
        return f"""# {prompt}

def solve_problem():
    '''
    Solution for: {prompt}
    '''
    # TODO: Implement your solution here
    pass

if __name__ == "__main__":
    solve_problem()"""
    elif language == "javascript":
        return f"""// {prompt}

function solveProblem() {{
    /*
    Solution for: {prompt}
    */
    // TODO: Implement your solution here
}}

solveProblem();"""
    else:
        return f"// {prompt}\n// TODO: Implement solution in {language}"

def simple_web_search(query: str, max_results: int = 5) -> List[Dict[str, Any]]:
    """Simple web search using DuckDuckGo instant answer API"""
    try:
        # DuckDuckGo instant answer API
        response = requests.get(
            "https://api.duckduckgo.com/",
            params={
                "q": query,
                "format": "json",
                "no_html": "1",
                "skip_disambig": "1"
            },
            timeout=10
        )
        
        if response.status_code == 200:
            data = response.json()
            results = []
            
            # Add abstract if available
            if data.get("AbstractText"):
                results.append({
                    "title": data.get("AbstractSource", "DuckDuckGo"),
                    "snippet": data.get("AbstractText"),
                    "url": data.get("AbstractURL", ""),
                    "type": "abstract"
                })
            
            # Add related topics
            for topic in data.get("RelatedTopics", [])[:max_results]:
                if isinstance(topic, dict) and topic.get("Text"):
                    results.append({
                        "title": topic.get("FirstURL", "").split("/")[-1] or "Related Topic",
                        "snippet": topic.get("Text"),
                        "url": topic.get("FirstURL", ""),
                        "type": "related"
                    })
            
            return results[:max_results]
        
        return []
    except Exception as e:
        return [{"title": "Search Error", "snippet": f"Could not perform search: {str(e)}", "url": "", "type": "error"}]

# Endpoints
@app.get("/")
async def root():
    return {
        "message": "Autonomous AI Agent API",
        "status": "active",
        "version": "1.0.0",
        "timestamp": datetime.now().isoformat()
    }

@app.get("/health")
async def health():
    return {
        "status": "healthy",
        "timestamp": datetime.now().isoformat(),
        "features": ["code_generation", "web_search", "reasoning"]
    }

@app.post("/generate")
async def generate_code(request: CodeRequest, token: bool = Depends(verify_token)):
    try:
        # Generate code using templates
        generated_code = generate_code_template(request.prompt, request.language)
        
        return {
            "code": generated_code,
            "explanation": f"Generated {request.language} code based on the pattern in your prompt: '{request.prompt}'",
            "confidence": 0.85,
            "language": request.language,
            "method": "template_based",
            "timestamp": datetime.now().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Code generation failed: {str(e)}")

@app.post("/search")
async def search_web(request: SearchRequest, token: bool = Depends(verify_token)):
    try:
        # Perform web search
        results = simple_web_search(request.query, request.max_results)
        
        return {
            "query": request.query,
            "results": results,
            "total_found": len(results),
            "summary": f"Found {len(results)} results for query: {request.query}",
            "timestamp": datetime.now().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Search failed: {str(e)}")

@app.post("/reason")
async def reason_problem(request: ReasonRequest, token: bool = Depends(verify_token)):
    try:
        problem = request.problem
        domain = request.domain
        
        # Simple reasoning steps
        reasoning_steps = [
            {
                "step": 1,
                "description": "Problem Analysis",
                "content": f"Analyzing the {domain} problem: {problem}"
            },
            {
                "step": 2,
                "description": "Solution Planning",
                "content": f"Breaking down the problem into manageable components"
            },
            {
                "step": 3,
                "description": "Implementation Strategy",
                "content": f"Determining the best approach for solving this {domain} challenge"
            }
        ]
        
        # Generate a basic solution outline
        if domain == "coding":
            solution = f"""
Approach for solving: {problem}

1. Understand the requirements
2. Design the algorithm/data structure
3. Implement the solution
4. Test with examples
5. Optimize if needed

This is a {domain} problem that can be solved systematically.
"""
        else:
            solution = f"Solution approach for {domain} problem: {problem}"
        
        return {
            "problem": problem,
            "domain": domain,
            "reasoning_steps": reasoning_steps,
            "solution": solution.strip(),
            "confidence": 0.75,
            "timestamp": datetime.now().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Reasoning failed: {str(e)}")

@app.get("/status")
async def get_status(token: bool = Depends(verify_token)):
    return {
        "status": "active",
        "features": [
            "template_based_code_generation",
            "web_search_via_duckduckgo",
            "step_by_step_reasoning",
            "jwt_authentication"
        ],
        "version": "1.0.0",
        "deployment": "vercel_optimized",
        "memory_usage": "minimal",
        "timestamp": datetime.now().isoformat()
    }

# Additional utility endpoints
@app.get("/features")
async def list_features():
    return {
        "code_generation": {
            "supported_languages": ["python", "javascript", "java", "cpp", "go"],
            "templates": ["fibonacci", "binary_search", "hello_world", "functions"],
            "method": "pattern_matching"
        },
        "web_search": {
            "provider": "duckduckgo",
            "features": ["instant_answers", "related_topics"],
            "rate_limit": "10_requests_per_minute"
        },
        "reasoning": {
            "domains": ["coding", "algorithms", "general"],
            "steps": "3_step_analysis"
        }
    }