requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Authentication and security
pyjwt==2.8.0

# Web scraping and search
requests==2.31.0

# Utilities
python-dotenv==1.0.0
